start fixing tests

+++ test_pyship/conftest.py
@@ -1,15 +1,16 @@
 import pytest
 from pathlib import Path
 import logging
-import sys
+import os
+import time
+from datetime import timedelta
 from semver import VersionInfo
+import subprocess
 
 from balsa import Balsa
 
-from pyship import mkdirs, subprocess_run, pyship_print, rmdir
-from pyship import __application_name__ as pyship_application_name
+from pyship import __application_name__ as pyship_application_name, rmdir
 from pyship import __author__ as pyship_author
-from pyship import __version__ as pyship_version
 
 from test_pyship import TST_APP_NAME, TstAppDirs
 
@@ -32,18 +33,21 @@ def session_fixture():
 
     balsa.init_logger()
 
-    # remove test app dirs
+    # create app environments
+
+    # even though the environments rarely change, recreate the environment only if it's older than this just to make sure we still can
+    env_timeout_sec = timedelta(hours=1).total_seconds()
+
+    subprocess.run(["build.bat"], shell=True)  # update pyship wheel to the latest
+
     for version_string in ["0.0.1", "0.0.2"]:
-        version = VersionInfo.parse(version_string)
-        tst_app_dirs = TstAppDirs(TST_APP_NAME, version)
+        tst_app_dirs = TstAppDirs(TST_APP_NAME, VersionInfo.parse(version_string))
+
         rmdir(tst_app_dirs.app_dir)
-        rmdir(tst_app_dirs.dist_dir)
 
-    # build pyship into a distributable package in the "dist" directory
-    # setup.py bdist_wheel
-    mkdirs(Path("dist"), remove_first=True)
-    subprocess_run([str(Path("venv", "Scripts", "python.exe")), "setup.py", "bdist_wheel"], mute_output=True, stderr_log=logging.info)  # flit writes output to stderr, not sure why
-    pyship_print(f"{pyship_application_name=} {pyship_version=} {pyship_author=}")
+        if os.path.getmtime(str(tst_app_dirs.venv_dir)) + env_timeout_sec < time.time():
+            rmdir(tst_app_dirs.dist_dir)
+            rmdir(tst_app_dirs.venv_dir)
 
-    if pyship_application_name not in sys.path:
-        sys.path.append(pyship_application_name)  # since, when testing pyship, pyship itself isn't in the venv
+            subprocess.run(["make_venv.bat"], cwd=tst_app_dirs.project_dir, shell=True)  # create the venv
+            subprocess.run(["build.bat"], cwd=tst_app_dirs.project_dir, shell=True)  # make the dist (wheel)
